requirements.txt
django==2.0
djangorestframework==3.7.7
psycopg2==2.7.3.2
flake8==3.5.0
django-ckeditor
fabric
tox
